create js to calculate total salary pretension page

+++ db/migrate/20180921151908_create_candidate_pretensions.rb
@@ -3,7 +3,7 @@ class CreateCandidatePretensions < ActiveRecord::Migration[5.1]
     create_table :candidate_pretensions do |t|
       t.references :candidate
       t.decimal :last_monthly_salary, :precision => 8, :scale => 2
-      t.integer :nofsalaries
+      t.decimal :nofsalaries
       t.decimal :variable, :precision => 8, :scale => 2
       t.decimal :last_salary_total, :precision => 8, :scale => 2
 			t.decimal :pretension_yearly, :precision => 8, :scale => 2
